feat(chat-core): add history helpers + 3 callbacks for read-only composers
